pyproject.toml
[tool.poetry]
name = "etlweather"
version = "0.1.0"
description = ""
authors = ["Jenil Shah <shahjenil000@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.9"
apache-airflow = "^2.10.4"
apache-airflow-providers-postgres = "^6.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
